+++ app/routes/extra_functions.py
@@ -1,0 +1,61 @@
+"""
+ Funciones extra
+"""
+
+
+#modulos externos
+from flask                                import Blueprint,redirect,flash,render_template
+from flask_login                          import current_user,login_required,logout_user
+
+#modulos propios
+from ..extra_functions.token              import confirm_token,genera_token
+from ..extra_functions.notification_funct import send_gmail_confirmation
+from ..extra_functions.email_decorator    import email_validation
+from ..models.importaciones               import User
+from ..controllers.importaciones          import UserController
+extra_functions = Blueprint("extra_functions",__name__)
+
+
+@extra_functions.route('/') 
+@extra_functions.route('/index') 
+def index(): 
+    if current_user.is_authenticated:
+        if current_user.email_conf is False:
+            return render_template("extra_functions/confirmation.html")
+        else:    
+            return redirect('/home')
+    else:
+        return redirect("/public/iniciar")
+                                       
+@extra_functions.route("/conf_email/<token>")
+@login_required
+def confirm_email(token):
+    if current_user.email_conf:
+        return redirect("/index")
+    email = confirm_token(token)
+    user = User().get_by_id(current_user.id)
+    if user.email == email:
+        user.email_conf= True
+        UserController().update_user(user)
+        flash("Cuenta confirmada")
+        return redirect("/index")
+    else:
+        flash("Token invalido")
+        #aqui añadir que muestre token vencido
+        return "xd"
+
+@extra_functions.route("/reenviartoken")
+@login_required
+def reenviar_token():
+    user  = User().get_by_id(current_user.id)
+    token = genera_token(user.email)
+    send_gmail_confirmation(token)
+    return redirect("https://mail.google.com/")
+
+
+@extra_functions.route("/cerrar_sesion")
+@login_required
+@email_validation
+def cerrar():
+    logout_user()
+    return redirect("/")
